pyproject.toml
[project]
name = "geosearch"
version = "0.1.0"
description = "Geospatial search tool for searching across multiple datasets"
url = "https://github.com/Amsterdam/geosearch/"
author = [
    { name = "Jasper Jager", email = "j.jager@amsterdam.nl" }
]
license = "EUPL-1.2"
requires-python = ">=3.14"
dependencies = [
    "adrf>=0.1.12",
    "azure-identity>=1.25.2",
    "azure-monitor-opentelemetry>=1.8.6",
    "datapunt-authorization-django>=2.0.2",
    "django==5.2.12",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.1.0",
    "psycopg>=3.3.3",
    "python-json-logger>=4.0.0",
    "uvicorn>=0.41.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "django-coverage-plugin>=3.2.0",
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "django-stubs>=5.2.9",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"

# ==== pytest ====
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
minversion = "6.0"
django_find_project = false
pythonpath = "."
testpaths = ["tests"]
addopts = [
    "--nomigrations",
    "--reuse-db",
    "--ds=tests.settings",
    "--cov",
    "--cov-fail-under=95"
]
norecursedirs = ["node_modules", ".tox", ".git"]
filterwarnings = [
    "once::DeprecationWarning",
    "once::PendingDeprecationWarning",
]
xfail_strict = true

# ==== Coverage ====
[tool.coverage.run]
branch = true
source = ["src/geosearch"]
omit = ["*/migrations/*.py", "settings.py", "asgi.py"]
plugins = ["django_coverage_plugin"]

[tool.coverage.report]
exclude_also = [
    "if settings.DEBUG:",
    'if "debug_toolbar" in settings.INSTALLED_APPS:',
    "raise NotImplementedError",
    "raise DomainException",
    "case exceptions.DomainException"
]

# ==== ruff ====
[tool.ruff]
line-length = 99
target-version = "py314"

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "I", # isort
    "B", # flake8-bugbear
    "C90", # mccabe
    "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "DJ", # flake8-django
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "RET", # flake8-return (partially)
    # "PT",  # flake8-pytest-style
    # "TCH",  # flake8-type-checking (moves import to `if typing.TYPE_CHECKING`)
    # "ERA",  # eradicate (commented out code)
    # "TRY",  # tryceratops
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "INT", # flake8-gettext
    "FLY", # flynt (fixes unneeded static string joins)
    "UP", # pyupgrade
    "S", # security (bugbear)
    "RUF010", # ruff: fix f"{str(..)}" usage
    "RUF013", # ruff: fix annotations for =None arguments
]
ignore = [
    "S311", # allow random.randint()
    "DJ001", # allow models.CharField(null=True)
    "SIM105", # enforcing contextlib.suppress() instead of try..catch
    "RET501", # unnecessary-return-none
    "RET505", # superfluous-else-return
    "RET505", # superfluous-else-return
    "RET506", # superfluous-else-raise
    "RET507", # superfluous-else-continue
    "RET508", # superfluous-else-break
    "S607", # subprocess partial path
]

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true

[tool.ruff.lint.flake8-gettext]
extend-function-names = ["gettext_lazy", "ngettext_lazy", "pgettext", "pgettext_lazy", "npgettext", "npgettext_lazy"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"django.utils.timezone.make_aware".msg = "There is no need for make_aware(), pass tzinfo directly."

[tool.ruff.lint.isort]
known-first-party = ["geosearch", "tests"]
#required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501"]  # line too long
"docs/_ext/djangodummy/settings.py" = ["S105"]  # allow hardcoded SECRET_KEY
"tests/settings.py" = ["F405"]  # allow unknown variables via import from *
"tests/**/*.py" = ["DJ008", "S101", "S105", "S106", "S314", "S320", "S608"]  # allow asserts, hardcoded passwords, lxml parsing, SQL injection
